+++ src/Synapse/Synapse/Commands/DataPlaneCommands/Artifact/Metastore/RegisterAzureSynapseMetastore.cs
@@ -1,0 +1,68 @@
+﻿// ----------------------------------------------------------------------------------
+//
+// Copyright Microsoft Corporation
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ----------------------------------------------------------------------------------
+
+using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
+using Microsoft.Azure.Commands.Synapse.Common;
+using Microsoft.Azure.Commands.Synapse.Models;
+using Microsoft.Azure.Commands.Synapse.Properties;
+using Microsoft.WindowsAzure.Commands.Utilities.Common;
+using System;
+using System.Management.Automation;
+
+namespace Microsoft.Azure.Commands.Synapse
+{
+    [Cmdlet(VerbsLifecycle.Register, ResourceManager.Common.AzureRMConstants.AzureRMPrefix + SynapseConstants.SynapsePrefix + SynapseConstants.Metastore,
+        DefaultParameterSetName = RegisterByName, SupportsShouldProcess = true)]
+    [OutputType(typeof(PSMetastore))]
+    public class RegisterAzureSynapseMetastore : SynapseArtifactsCmdletBase
+    {
+        private const string RegisterByName = "RegisterByName";
+        private const string RegisterByObject = "RegisterByObject";
+
+        [Parameter(ValueFromPipelineByPropertyName = false, ParameterSetName = RegisterByName,
+            Mandatory = true, HelpMessage = HelpMessages.WorkspaceName)]
+        [ResourceNameCompleter(ResourceTypes.Workspace, "ResourceGroupName")]
+        [ValidateNotNullOrEmpty]
+        public override string WorkspaceName { get; set; }
+
+        [Parameter(ValueFromPipeline = true, ParameterSetName = RegisterByObject,
+            Mandatory = true, HelpMessage = HelpMessages.WorkspaceObject)]
+        [ValidateNotNull]
+        public PSSynapseWorkspace WorkspaceObject { get; set; }
+
+        [Parameter(ValueFromPipelineByPropertyName = false, Mandatory = true, HelpMessage = HelpMessages.CreatedDatabaseName)]
+        [ValidateNotNullOrEmpty]
+        public string DatabaseName { get; set; }
+
+        [Parameter(ValueFromPipelineByPropertyName = false, Mandatory = true, HelpMessage = HelpMessages.InputFolder)]
+        [ValidateNotNullOrEmpty]
+        public string InputFolder { get; set; }
+
+        [Parameter(Mandatory = false, HelpMessage = HelpMessages.AsJob)]
+        public SwitchParameter AsJob { get; set; }
+
+        public override void ExecuteCmdlet()
+        {
+            if (this.IsParameterBound(c => c.WorkspaceObject))
+            {
+                this.WorkspaceName = this.WorkspaceObject.Name;
+            }
+
+            if (this.ShouldProcess(this.WorkspaceName, String.Format(Resources.RegisteringSynapseMetastore, this.DatabaseName, this.WorkspaceName)))
+            {
+                WriteObject(new PSMetastore(SynapseAnalyticsClient.RegisterMetaStore(this.DatabaseName, this.InputFolder), this.WorkspaceName, this.DatabaseName));
+            }
+        }
+    }
+}