+++ src/Synapse/Synapse/Common/HelpMessages.cs
@@ -562,5 +562,11 @@ For more information, see https://docs.microsoft.com/en-us/sql/relational-databa
         public const string KustoPoolDatabaseName = "Name of Synapse Kusto database.";
 
         public const string StorageAccountType = "The storage account type used to store backups for the sql pool. Possible values include: 'GRS', 'LRS'.";
+
+        public const string CreatedDatabaseName = "The name of the database to be created. The name can contain only alphanumeric characters and should not exceed 24 characters.";
+
+        public const string DatabaseName = "The name of the database.";
+
+        public const string InputFolder = "The input folder containing CDM files.";
     }
 }