+++ src/Accounts/Accounts/ChangeLog.md
@@ -22,6 +22,7 @@
 * Enabled Continue Access Evaluation for MSGraph
 * Improved error message when login is blocked by AAD
 * Improved error message when silent reauthentication failed
+* Upgraded Azure.Core from 1.21.0 to 1.22.0.
 
 ## Version 2.7.2
 * Removed legacy assembly System.Private.ServiceModel and System.ServiceModel.Primitives [#16063]